Create Final
Sahr al-Sutur” (Magic of the Lines) is an interactive project aimed at encouraging children to develop a love for reading through a magical story about books. The project follows an exciting adventure where the main character discovers the magic of books and their ability to transport them to new, fascinating worlds, promoting the importance of reading as a means of expanding knowledge and imagination.

The project uses modern animation techniques (such as Adobe After Effects and Blender) to create visually engaging content designed to capture children’s attention and motivate them to read. Characters and environments are designed interactively, allowing children to engage with the story and learn from it. Additionally, music and sound effects are added to enhance the experience and make it more enjoyable.

The goal of the project is to create a fun, educational environment that encourages children to explore the world of reading in an innovative artistic way, helping them develop language skills and critical thinking through an entertaining and immersive experience.

Signed-off-by: yuliana-magdy <magdymikhailbushrayuliana@gmail.com>